requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
djangorestframework==3.10.3
drf-url-filters==0.5.1
idna==2.8
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6
voluptuous==0.11.7
